// src/apis/generated/models/snsInfoPlatform.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Project Name
 * 개발: [api.dev.my-domain.com](https://api.dev.my-domain.com)
<br/>
운영: [api.my-domain.com](https://api.my-domain.com)

 * OpenAPI spec version: 1.0.0
 */

export type SnsInfoPlatform =
  (typeof SnsInfoPlatform)[keyof typeof SnsInfoPlatform]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SnsInfoPlatform = {
  INSTAGRAM: 'INSTAGRAM',
  TWITTER: 'TWITTER',
  FACEBOOK: 'FACEBOOK',
  LINKEDIN: 'LINKEDIN',
  OTHERS: 'OTHERS',
} as const